Gcpedia/python_requirements.txt
django-bootstrap-static
